# openapi.yaml
openapi: 3.0.3
info:
  title: ACS Multi-Vendor TR-069/CWMP REST API
  version: 1.0.0
  description: |
    REST API internal (Echo v5) untuk Auto Configuration Server (ACS) TR-069/CWMP
    multi-vendor (ZTE, Huawei, FiberHome, Nokia, C-Data, dst). Dikonsumsi oleh
    sistem BSS/OSS, portal NOC, dan integrasi internal lain — **BUKAN** endpoint
    yang dipakai CPE (ONT/ONU/router) itu sendiri.

    **PENTING — beda dengan endpoint CWMP:** perangkat CPE berkomunikasi dengan
    ACS lewat endpoint `/cwmp` (SOAP/XML, protokol TR-069/CWMP standar Broadband
    Forum, sesi stateful) yang **TIDAK** didokumentasikan di file ini sama
    sekali — beda protokol, beda tujuan, beda konsumen. Dokumen ini HANYA
    mencakup `/api/v1/*` (REST + JSON, kecuali `GET /api/v1/metrics` yang
    formatnya teks Prometheus).

    ## Autentikasi

    Hampir seluruh endpoint mewajibkan header `Authorization: Bearer <token>`,
    berupa salah satu dari:
    - **JWT sesi user** — didapat dari `POST /auth/login`, berlaku sesuai masa
      berlaku yang dikonfigurasi (`ACS_JWT_EXPIRY`, default 8 jam di `.env.example`).
    - **API token** — string berawalan `acs_...`, diterbitkan lewat
      `POST /auth/tokens` (ADMIN/SUPERADMIN), tidak pernah kedaluwarsa kecuali
      `expires_at` diisi.

    Empat endpoint publik tanpa autentikasi, semuanya READ-ONLY atau titik masuk
    login (tidak ada endpoint MUTASI tanpa autentikasi): `POST /auth/login`
...
    aman).

...

    ## RBAC

...

...

...
    skema `Task`.

    ## Server

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [Catalog]
      summary: Daftar vendor CPE
      description: |
        Semua role terautentikasi. Data referensi GLOBAL lintas tenant (FR-28)
        — TIDAK ada filter tenant sama sekali di endpoint katalog vendor/model/
        mapping (by design, ini "data referensi platform" bukan data tenant).
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          description: Daftar vendor.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VendorListResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    post:
      operationId: createVendor
      tags: [Catalog]
      summary: Menambah vendor CPE baru
      description: |
        Superadmin only. Menambah dukungan vendor baru adalah OPERASI DATA
        (insert ke `ref_vendors`/`vendor_ouis`/`device_models`/
        `vendor_parameter_mappings`), bukan perubahan kode — lihat juga
        `POST /vendors/:id/ouis`, `POST /device-models`,
        `POST /vendor-parameter-mappings`.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateVendorRequest' }
      responses:
        '201':
          description: Vendor dibuat.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vendor' }
        '400':
          description: "Payload tidak valid, `code`/`name` kosong."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Bukan superadmin.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: "`code` vendor sudah dipakai."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /vendors/{id}/ouis:
    get:
      operationId: listVendorOUIs
      tags: [Catalog]
      summary: Daftar OUI (prefix MAC) terdaftar milik vendor
      description: Semua role terautentikasi. Data global lintas tenant.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/vendorIdPath'
      responses:
        '200':
          description: Daftar OUI (array polos, TIDAK dibungkus `data`/`total`).
          content:
            application/json:
              schema:
                type: array
                nullable: true
                items: { $ref: '#/components/schemas/VendorOUI' }
        '400':
          description: "`id` tidak valid."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    post:
      operationId: addVendorOUI
      tags: [Catalog]
      summary: Mendaftarkan OUI baru milik vendor
      description: |
        Superadmin only. OUI dipakai me-resolve vendor device baru saat Inform
        pertama kali (lihat `FindOrCreateFromInform`).
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/vendorIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddVendorOUIRequest' }
      responses:
        '201':
          description: OUI ditambahkan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VendorOUI' }
        '400':
          description: "`id` tidak valid, payload tidak valid, atau `oui` kosong."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Bukan superadmin.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: OUI sudah terdaftar.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /device-models:
    get:
      operationId: listDeviceModels
      tags: [Catalog]
      summary: Daftar model device milik satu vendor
      description: Semua role terautentikasi. Data global lintas tenant. `vendor_id` WAJIB diisi.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vendor_id
          in: query
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Daftar device model (array polos, TIDAK dibungkus `data`/`total`).
          content:
            application/json:
              schema:
                type: array
                nullable: true
                items: { $ref: '#/components/schemas/DeviceModel' }
        '400':
          description: "`vendor_id` tidak diisi."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    post:
      operationId: createDeviceModel
      tags: [Catalog]
      summary: Menambah model device baru milik vendor
      description: Superadmin only — bagian dari operasi data menambah dukungan vendor/model (lihat `POST /vendors`).
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDeviceModelRequest' }
      responses:
        '201':
          description: Device model dibuat.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceModel' }
        '400':
          description: "Payload tidak valid, `vendor_id`/`model_name` kosong."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Bukan superadmin.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT sesi user (dari `POST /auth/login`) ATAU API token statis
        berawalan `acs_...` (dari `POST /auth/tokens`) — keduanya dikirim
        dengan cara yang sama lewat header `Authorization: Bearer <token>`.
  parameters:
    page:
      name: page
      in: query
      description: Halaman (1-based). Default 1.
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      name: page_size
      in: query
      description: Ukuran halaman. Default 50, maksimum 200 (nilai di luar rentang ini jatuh ke default 50).
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
    vendorIdPath:
      name: id
      in: path
      required: true
      description: ID numerik internal vendor.
      schema: { type: integer, format: int64 }
...
  schemas:
...
    ErrorResponse:
      type: object
      description: |
        Bentuk error SERAGAM untuk seluruh endpoint (`echo.NewHTTPError`) —
        selalu `{"message": "..."}`, tidak ada kode error mesin terpisah dari
        `message` teks bebas saat ini (perhatikan: teks pesan sebagian besar
        Bahasa Indonesia, konsisten dengan gaya pesan aplikasi ini).
      properties:
        message:
          type: string
          example: "domain: data tidak ditemukan"
      required: [message]
...
    Audit:
      type: object
      description: 7 kolom audit standar yang dipakai hampir semua entitas master/transactional.
      properties:
        created_at: { type: string, format: date-time }
        created_by: { type: integer, format: int64, nullable: true }
        updated_at: { type: string, format: date-time }
        updated_by: { type: integer, format: int64, nullable: true }
        deleted_at: { type: string, format: date-time, nullable: true }
        deleted_by: { type: integer, format: int64, nullable: true }
        is_deleted: { type: boolean }

    # ---------- Auth ----------
...
    Vendor:
      type: object
      properties:
        id: { type: integer, format: int64 }
        code: { type: string, example: ZTE }
        name: { type: string, example: ZTE Corporation }
        description: { type: string, nullable: true }
        is_active: { type: boolean }
      allOf:
        - $ref: '#/components/schemas/Audit'
      required: [id, code, name, is_active]
    VendorListResponse:
      type: object
      properties:
        data:
          type: array
          nullable: true
          items: { $ref: '#/components/schemas/Vendor' }
        total: { type: integer }
      required: [data, total]
    CreateVendorRequest:
      type: object
      properties:
        code: { type: string, maxLength: 32 }
        name: { type: string, maxLength: 128 }
        description: { type: string, nullable: true, maxLength: 255 }
      required: [code, name]
    VendorOUI:
      type: object
      properties:
        id: { type: integer, format: int64 }
        vendor_id: { type: integer, format: int64 }
        oui: { type: string, example: "AABBCC" }
        notes: { type: string, nullable: true }
      allOf:
        - $ref: '#/components/schemas/Audit'
      required: [id, vendor_id, oui]
    AddVendorOUIRequest:
      type: object
      properties:
        oui:
          type: string
          maxLength: 6
          description: |
            Kolom `vendor_ouis.oui` CHAR(6). Server HANYA menegakkan batas ATAS
            (maks 6 karakter) — nilai lebih pendek TETAP diterima, dan tidak ada
            validasi "wajib 6 digit heksadesimal" (keputusan produk eksplisit di
            `delivery/http/validate.go`, bukan kelalaian).
          example: "AABBCC"
        notes: { type: string, nullable: true, maxLength: 255 }
      required: [oui]
    DeviceModel:
      type: object
      properties:
        id: { type: integer, format: int64 }
        vendor_id: { type: integer, format: int64 }
        device_type_id: { type: integer, format: int64 }
        data_model_version_id: { type: integer, format: int64, description: "Resolve ke `InternetGatewayDevice.*` (TR-098) atau `Device.*` (TR-181)." }
        product_class: { type: string, nullable: true }
        model_name: { type: string }
        description: { type: string, nullable: true }
        is_active: { type: boolean }
      allOf:
        - $ref: '#/components/schemas/Audit'
      required: [id, vendor_id, device_type_id, data_model_version_id, model_name, is_active]
    CreateDeviceModelRequest:
      type: object
      properties:
        vendor_id: { type: integer, format: int64 }
        device_type_id: { type: integer, format: int64 }
        data_model_version_id: { type: integer, format: int64 }
        product_class: { type: string, nullable: true, maxLength: 128 }
        model_name: { type: string, maxLength: 128 }
        description: { type: string, nullable: true, maxLength: 255 }
      required: [vendor_id, device_type_id, data_model_version_id, model_name]
...